+++ src/main/kotlin/com/example/days/domain/admin/controller/AdminController.kt
@@ -5,7 +5,11 @@ import com.example.days.domain.admin.dto.request.SignUpAdminRequest
 import com.example.days.domain.admin.dto.response.AdminResponse
 import com.example.days.domain.admin.dto.response.LoginAdminResponse
 import com.example.days.domain.admin.service.AdminService
+import com.example.days.domain.messages.dto.request.CreateMessageRequest
+import com.example.days.domain.messages.dto.response.AdminMessagesSendResponse
+import com.example.days.domain.report.dto.response.UserReportResponse
 import com.example.days.domain.user.dto.response.UserResponse
+import com.example.days.global.infra.security.UserPrincipal
 import io.swagger.v3.oas.annotations.Operation
 import jakarta.validation.Valid
 import org.springframework.data.domain.Page
@@ -14,6 +18,7 @@ import org.springframework.data.web.PageableDefault
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.security.access.prepost.PreAuthorize
+import org.springframework.security.core.annotation.AuthenticationPrincipal
 import org.springframework.web.bind.annotation.*
 
 @RestController
@@ -74,6 +79,39 @@ class AdminController(
     ): ResponseEntity<Unit> {
         adminService.adminBanByAdmin(adminId)
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build()
+    }
+
+    @Operation(summary = "신고 당한 유저 조회")
+    @PreAuthorize("hasRole('ADMIN')")
+    @GetMapping("/reports/users")
+    fun getReportUser(
+        @PageableDefault(size = 10, sort = ["nickname"]) pageable: Pageable,
+        @RequestParam nickname: String,
+        @AuthenticationPrincipal userPrincipal: UserPrincipal
+    ): ResponseEntity<Page<UserReportResponse>> {
+        return ResponseEntity.status(HttpStatus.OK).body(adminService.getReportUser(pageable, nickname))
+    }
+
+    @Operation(summary = "TO 유저 By 어드민 쪽지 생성")
+    @PreAuthorize("hasRole('ADMIN')")
+    @PostMapping("/messages")
+    fun toUserCreateMessage(
+        @RequestBody req: CreateMessageRequest,
+        @AuthenticationPrincipal userPrincipal: UserPrincipal
+    ): ResponseEntity<AdminMessagesSendResponse> {
+        val userId = userPrincipal.id
+        return ResponseEntity.status(HttpStatus.CREATED).body(adminService.toUserCreateMessage(req, userId))
+    }
+
+    @Operation(summary = "어드민 쪽지 전체 조회 Only 어드민만")
+    @PreAuthorize("hasRole('ADMIN')")
+    @GetMapping("/messages/read/all")
+    fun readAllMessagesOnlyAdmin(
+        @PageableDefault(size = 10, sort = ["sentAt"]) pageable: Pageable,
+        @AuthenticationPrincipal userPrincipal: UserPrincipal
+    ): ResponseEntity<Page<AdminMessagesSendResponse>> {
+        val userId = userPrincipal.id
+        return ResponseEntity.status(HttpStatus.OK).body(adminService.readAllMessagesOnlyAdmin(pageable, userId))
     }
 
 }
