+++ src/main/kotlin/com/example/days/domain/admin/service/AdminService.kt
@@ -4,6 +4,9 @@ import com.example.days.domain.admin.dto.request.LoginAdminRequest
 import com.example.days.domain.admin.dto.request.SignUpAdminRequest
 import com.example.days.domain.admin.dto.response.AdminResponse
 import com.example.days.domain.admin.dto.response.LoginAdminResponse
+import com.example.days.domain.messages.dto.request.CreateMessageRequest
+import com.example.days.domain.messages.dto.response.AdminMessagesSendResponse
+import com.example.days.domain.report.dto.response.UserReportResponse
 import com.example.days.domain.user.dto.response.UserResponse
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
@@ -26,5 +29,14 @@ interface AdminService {
 
     // 어드민 밴 기능
     fun adminBanByAdmin(adminId: Long)
+
+    // 유저 신고 기능
+    fun getReportUser(pageable: Pageable, nickname: String): Page<UserReportResponse>
+
+    // 어드민이 유저에게 쪽지 보내는 기능
+    fun toUserCreateMessage(req: CreateMessageRequest, userId: Long) : AdminMessagesSendResponse
+
+    //오직 어드민이 보낸 쪽지 전체 조회 기능
+    fun readAllMessagesOnlyAdmin(pageable: Pageable, userId: Long): Page<AdminMessagesSendResponse>
 
 }
